changed: config.h will be always pre-included to set _FILE_OFFSET_BITS=64 and define off_t correctly in *nix
r3334

+++ FreeBASIC/src/rtlib/linux/fb_linux.h
@@ -149,7 +149,7 @@ typedef struct FBCONSOLE
 	void (*gfx_exit)(void);
 } FBCONSOLE;
 
-typedef long long int fb_off_t;
+typedef off_t fb_off_t;
 
 extern FBCONSOLE __fb_con;
